Rename theme and add file to override some css rules of storybook
