+++ scss/_storybook-overrides.scss
@@ -1,0 +1,10 @@
+@import url('https://fonts.googleapis.com/css?family=Merriweather:300,400,700,900&display=swap');
+
+.sb-show-main {
+  margin: 0 !important;
+}
+
+h1, h2, h3, h4, h5, p{
+color: red !important;
+font-family: 'Merriweather', serif !important;
+}
