+++ scss/main.scss
@@ -6,5 +6,6 @@ body {
   font-size: 1.6rem;
 }
 
+@import 'storybook-overrides';
 @import 'colors';
 @import 'example-button';
